BlueGreen/Version2/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async==0.6.2
asyncio==3.4.3
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dill==0.4.0
Faker==37.1.0
fastapi==0.115.12
flasgger==0.9.7.1
Flask==3.1.0
flask-marshmallow==1.3.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
marshmallow==3.26.1
mistune==3.1.3
multidict==6.4.4
multiprocess==0.70.18
packaging==24.2
pact-python==2.3.1
pika==1.3.2
pluggy==1.6.0
propcache==0.3.1
psutil==6.1.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pytest==8.3.5
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==3.1.3
yarl==1.20.0
zipp==3.21.0